+++ KITTexe.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-cd /opt/KITT
-python3 /opt/KITT/KITT2.py
-